vbm-auth/requirements.txt
bcrypt==4.0.1
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.11.7
pydantic-settings==2.9.1
motor==3.7.0
aiosmtplib==4.0.0
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.5

email_validator==2.2.0
starlette==0.46.1
